+++ Engine/src/frontend/Physics.h
@@ -24,10 +24,7 @@ namespace ph{
 				glm::vec3 position, glm::vec3 scale, float static_friction, float dynamic_friction, float restitution
 			) noexcept -> StaticCollider {
 
-				return this->interface->create_static_cube(
-					(alias::Vec3)&position, (alias::Vec3)&scale, static_friction, dynamic_friction, restitution
-				);
-
+				return this->interface->create_static_cube(position, scale, static_friction, dynamic_friction, restitution);
 			};
 
 
@@ -40,10 +37,7 @@ namespace ph{
 				glm::vec3 position, glm::vec3 scale, float static_friction, float dynamic_friction, float restitution
 			) noexcept -> DynamicCollider {
 
-				return this->interface->create_dynamic_cube(
-					(alias::Vec3)&position, (alias::Vec3)&scale, static_friction, dynamic_friction, restitution
-				);
-
+				return this->interface->create_dynamic_cube(position, scale, static_friction, dynamic_friction, restitution);
 			};
 
 
@@ -60,6 +54,25 @@ namespace ph{
 				return this->interface->get_dynamic_collider_transform(collider);
 			};
 
+
+
+			///////////////////////////////////
+			// character controller
+
+
+			PH_NODISCARD inline auto createCharacterController(glm::vec3 position, float height, float radius) noexcept -> CharacterController {
+				return this->interface->create_character_controller(position, height, radius);
+			};
+
+
+			PH_NODISCARD inline auto getCharacterControllerPosition(CharacterController controller) noexcept -> glm::vec3 {
+				return this->interface->get_character_controller_position(controller);
+			};
+
+
+			PH_NODISCARD inline auto characterControllerMove(CharacterController controller, glm::vec3 direction, float dt) noexcept -> void {
+				return this->interface->character_controller_move(controller, direction, dt);
+			};